til_agent/requirements.txt
google-adk>=0.5.0
google-cloud-alloydb-connector[pg8000]
pg8000
google-cloud-aiplatform
python-dotenv
